LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-coop-navigated-opener.https-expected.txt


FAIL navigation-report-from-opener-navigation assert_equals: expected (string) "coop" but got (undefined) undefined
